Moved the server config.js to the server folder.
