+++ node_server/server.js
@@ -1,6 +1,6 @@
 const http = require('http');
 
-const config = require('../config.js');
+const config = require('./config.js');
 const serverConfig = config();
 
 //mock data for testing
